+++ lib/regent.dart
@@ -2,7 +2,6 @@ export 'dart:async' show StreamSubscription;
 
 export 'package:identifiable/identifiable.dart';
 
-export 'src/envelope.dart';
 export 'src/covered_ranges.dart';
 export 'src/event_streams.dart';
 export 'src/guard.dart';